requirements.txt
# Core AI
torch>=2.2.0
torchvision
transformers>=4.35.0
diffusers>=0.24.0
accelerate>=0.26.0
safetensors

# Image Processing
opencv-python-headless
Pillow
numpy<2.0

# Backend & Queue
fastapi
uvicorn[standard]
python-multipart
redis
rq

# UI
streamlit

# Utilities
python-dotenv